pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coffer-config"
version = "0.1.0"
description = "Typed config reader for resolved Coffer config files"
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov"]

[tool.hatch.build.targets.wheel]
packages = ["src/coffer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
